refactor: clean cyclical dependencies
